Move probe-related types to "types" package.
