+++ packages/breadboard/src/harness/diagnostics.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { Probe, ProbeMessage } from "../types.js";
+import type { ProbeMessage } from "@breadboard-ai/types";
+import type { Probe } from "../types.js";
 
 export type DiagnosticsCallback = (message: ProbeMessage) => Promise<void>;
