+++ packages/breadboard/src/run/run-graph.ts
@@ -4,17 +4,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { GraphDescriptor, OutputValues } from "@breadboard-ai/types";
+import type {
+  GraphDescriptor,
+  OutputValues,
+  TraversalResult,
+} from "@breadboard-ai/types";
 import { bubbleUpInputsIfNeeded, bubbleUpOutputsIfNeeded } from "../bubble.js";
 import { resolveBoardCapabilities } from "../capability.js";
 import { InputStageResult, OutputStageResult } from "../run.js";
 import { timestamp } from "../timestamp.js";
 import { TraversalMachine } from "../traversal/machine.js";
-import type {
-  BreadboardRunResult,
-  RunArguments,
-  TraversalResult,
-} from "../types.js";
+import type { BreadboardRunResult, RunArguments } from "../types.js";
 import { asyncGen } from "../utils/async-gen.js";
 import { NodeInvoker } from "./node-invoker.js";
 import { cloneState } from "../serialization.js";
