+++ packages/breadboard/src/traversal/iterator.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { Edge, TraversalResult, ErrorCapability } from "../types.js";
+import type { Edge, TraversalResult } from "@breadboard-ai/types";
+import type { ErrorCapability } from "../types.js";
 import { Traversal } from "./index.js";
 import { GraphRepresentation } from "./representation.js";
 import { MachineResult } from "./result.js";
