+++ packages/breadboard/tests/inspector/inline-data-run.ts
@@ -205,7 +205,6 @@ export const results: HarnessRunResult[] = [
           ],
         },
       },
-      validatorMetadata: [],
       path: [1],
       timestamp: 16784.899999976158,
       newOpportunities: [],
@@ -352,7 +351,6 @@ export const results: HarnessRunResult[] = [
           ],
         },
       },
-      validatorMetadata: [],
       path: [2],
       timestamp: 16803.69999998808,
       outputs: {},
