+++ packages/types/src/index.ts
@@ -6,3 +6,5 @@
 
 export type * from "./graph-descriptor.js";
 export type * from "./llm-content.js";
+export type * from "./probe.js";
+export type * from "./traversal.js";
